[Editorial] Refresh xref usage, drop custom script
See discussion in https://github.com/w3c/media-source/pull/337#issuecomment-1837837578

The references to ID3v1 and ID3v2 remain in the local biblio for now because
the id3.org server, which hosts the specifications, returns a 500 Internal
Server Error (and has been returning that error in the past few months). Not
quite sure what to do with that.

Also, Specref has ID3 v2.4.0, but the spec probably means it when it references
v2.3.0 as that is the "main" one in practice.